requirements.txt
# LLM Graph
langgraph>=0.2.0
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# Embeddings (local, free)
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0

# Vector store
faiss-cpu>=1.7.4

# Speech to text (local Whisper, GPU-accelerated)
openai-whisper>=20231117

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.0

# MCP Server
mcp[cli]>=1.0.0

# Cache
redis>=5.0.0

# Monitoring
prometheus-fastapi-instrumentator>=6.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
